refactor: move PokemonTypeChip, PokemonTypeChipGroup into ui.component
